requirements.txt
streamlit == 1.39.0
pydantic-settings == 2.5.2
